refactor(frontend): organize webapi calls in a single module
