+++ frontend/src/upnpapi/api.ts
@@ -1,4 +1,5 @@
 import { PlaybackInfo, PlayerDevice } from "./types";
+import { LibraryListItem } from "./types";
 import { adaptTo } from "./utils";
 
 export function getDevices() {
@@ -20,7 +21,7 @@ export function play(playerId: string, itemid: string) {
   });
 }
 
-export function playbackInfo(playerId: string) {
+export function getPlaybackInfo(playerId: string) {
   const url = `/api/player/${playerId}/playback`;
   return fetch(url)
     .then((response: any) => response.json())
@@ -39,4 +40,21 @@ export function setVolume(playerId: string, volume: number) {
     },
     body: JSON.stringify({ volume_percent: volume }),
   });
+}
+
+export function browse(id?: string) {
+  let url = `/api/library/`;
+  if (id) {
+    url = `/api/library/${id}`;
+  }
+  return fetch(url)
+    .then((response: any) => response.json())
+    .then((data: any) => adaptTo<LibraryListItem[]>(data));
+}
+
+export function getItem(id: string) {
+  const url = `/api/library/${id}/metadata`;
+  return fetch(url)
+    .then((response: any) => response.json())
+    .then((data: any) => adaptTo<LibraryListItem>(data));
 }
